Migrated Config to common (#607)
* Updated to camel case.

* Fixed formatting.

* to reflect upstream develop

* got rid of regex and changed formatting of str names

* pulled recent changes from upstream develop

* migrated config to common

* fixed lint issues

* Corrected all the changes

* syntax err fixed

* Fixed import

* Latest upstream changes pulled

* Fixed Linting errors

* Lint fix

* Pulled latest

* fixed other occurences
