+++ src/weathergen/model/model.py
@@ -20,6 +20,7 @@ import torch
 from astropy_healpix import healpy
 from torch.utils.checkpoint import checkpoint
 
+from weathergen.common.config import Config
 from weathergen.model.engines import (
     EmbeddingEngine,
     EnsPredictionHead,
@@ -33,8 +34,8 @@ from weathergen.model.engines import (
 from weathergen.model.layers import MLP, NamedLinear
 from weathergen.model.parametrised_prob_dist import LatentInterpolator
 from weathergen.model.utils import get_num_parameters
-from weathergen.utils.config import Config, get_dtype
 from weathergen.utils.logger import logger
+from weathergen.utils.utils import get_dtype
 
 
 class ModelParams(torch.nn.Module):
