+++ src/weathergen/train/trainer_base.py
@@ -18,8 +18,8 @@ import torch
 import torch.distributed as dist
 import torch.multiprocessing
 
+from weathergen.common.config import Config
 from weathergen.train.utils import str_to_tensor, tensor_to_str
-from weathergen.utils.config import Config
 from weathergen.utils.distributed import is_root
 
 _logger = logging.getLogger(__name__)
