+++ src/weathergen/utils/train_logger.py
@@ -21,7 +21,7 @@ import numpy as np
 import polars as pl
 from torch import Tensor
 
-import weathergen.utils.config as config
+import weathergen.common.config as config
 from weathergen.utils.metrics import get_train_metrics_path, read_metrics_file
 
 _weathergen_timestamp = "weathergen.timestamp"
